rearrange exporters, rename package
