+++ src/main/scala/org/clulab/wm/eidos/apps/AnnotationTSV.scala
@@ -7,11 +7,11 @@ import org.clulab.odin.{EventMention, State}
 import org.clulab.struct.Counter
 import org.clulab.utils.Configured
 import org.clulab.wm.eidos.EidosSystem
-import org.clulab.wm.eidos.export.EntityInfo
+import org.clulab.wm.eidos.exporters._
 import org.clulab.wm.eidos.document.AnnotatedDocument
 import org.clulab.wm.eidos.mentions.EidosEventMention
 import org.clulab.wm.eidos.utils.Closer.AutoCloser
-import org.clulab.wm.eidos.utils.{ExportUtils, FileUtils}
+import org.clulab.wm.eidos.utils.FileUtils
 import ai.lum.common.StringUtils._
 
 object AnnotationTSV extends App with Configured {
@@ -47,7 +47,7 @@ object AnnotationTSV extends App with Configured {
       trigger = mention.odinMention.asInstanceOf[EventMention].trigger
       relation_txt = trigger.text.normalizeSpace
       relation_norm = mention.label // i.e., "Causal" or "Correlation"
-      relation_modifier = ExportUtils.getModifier(mention) // prob none
+      relation_modifier = getModifier(mention) // prob none
 
       effect <- mention.asInstanceOf[EidosEventMention].eidosArguments("effect")
       factor_b_info = EntityInfo(effect, groundAs)
