+++ src/main/scala/org/clulab/wm/eidos/apps/ExtractAndExport.scala
@@ -4,8 +4,9 @@ package org.clulab.wm.eidos.apps
 import com.typesafe.config.{Config, ConfigFactory}
 import org.clulab.utils.Configured
 import org.clulab.wm.eidos.EidosSystem
+import org.clulab.wm.eidos.exporters.getExporter
 import org.clulab.wm.eidos.groundings.CompositionalGrounder
-import org.clulab.wm.eidos.utils.{ExportUtils, FileUtils}
+import org.clulab.wm.eidos.utils.FileUtils
 
 /**
   * App used to extract mentions from files in a directory and produce the desired output format (i.e., jsonld, mitre
@@ -46,7 +47,7 @@ object ExtractAndExport extends App with Configured {
     val annotatedDocuments = Seq(reader.extractFromText(text, idOpt = Some(file.getName)))
     // 4. Export to all desired formats
     exportAs.foreach { format =>
-      ExportUtils.getExporter(format, s"$outputDir/${file.getName}", reader, groundedAs, topN).export(annotatedDocuments)
+      getExporter(format, s"$outputDir/${file.getName}", reader, groundedAs, topN).export(annotatedDocuments)
     }
   }
 }
