+++ src/main/scala/org/clulab/wm/eidos/exporters/Exporter.scala
@@ -1,0 +1,84 @@
+package org.clulab.wm.eidos.exporters
+
+import java.io.PrintWriter
+
+import org.clulab.odin.State
+import org.clulab.wm.eidos.EidosSystem
+import org.clulab.wm.eidos.document.AnnotatedDocument
+import org.clulab.wm.eidos.mentions.EidosEventMention
+import org.clulab.wm.eidos.utils.{FileUtils, MentionUtils}
+import org.clulab.wm.eidos.utils.Closer.AutoCloser
+
+
+trait Exporter {
+  def export(annotatedDocuments: Seq[AnnotatedDocument]): Unit
+}
+
+
+@deprecated("MitreExporter will be deprecated", "0.3.0")
+case class MitreExporter(outFilename: String, reader: EidosSystem, filename: String, groundAs: Seq[String], topN: Int) extends Exporter {
+
+  override def export(annotatedDocuments: Seq[AnnotatedDocument]): Unit = {
+    FileUtils.printWriterFromFile(outFilename).autoClose { pw =>
+      pw.println(header())
+      annotatedDocuments.foreach(printTableRows(_, pw, filename, reader))
+    }
+  }
+
+  def header(): String = {
+    val factors = Seq("A", "B").flatMap { factor =>
+      groundAs.map { namespace =>
+        s"Factor $factor top${topN}_${namespace.toUpperCase}Ontology"
+      }
+    }.mkString("\t")
+
+    "Source\tSystem\tSentence ID\tFactor A Text\tFactor A Normalization\t" +
+      "Factor A Modifiers\tFactor A Polarity\tRelation Text\tRelation Normalization\t" +
+      "Relation Modifiers\tFactor B Text\tFactor B Normalization\tFactor B Modifiers\t" +
+      "Factor B Polarity\tLocation\tTime\tEvidence\t" +
+      factors
+  }
+
+  def printTableRows(annotatedDocument: AnnotatedDocument, pw: PrintWriter, filename: String, reader: EidosSystem): Unit = {
+    val allOdinMentions = annotatedDocument.eidosMentions.map(_.odinMention)
+    val mentionsToPrint = annotatedDocument.eidosMentions.filter(m => reader.components.stopwordManager.releventEdge(m.odinMention, State(allOdinMentions)))
+
+    for {
+      mention <- mentionsToPrint
+
+      source = filename
+      system = "Eidos"
+      sentence_id = mention.odinMention.sentence
+
+      // For now, only put EidosEventMentions in the mitre tsv
+      if mention.isInstanceOf[EidosEventMention]
+      cause <- mention.asInstanceOf[EidosEventMention].eidosArguments("cause")
+      factor_a_info = EntityInfo(cause, groundAs)
+
+      trigger = MentionUtils.triggerOpt(mention).getOrElse("")
+      relation_txt = removeTabAndNewline(trigger)
+      relation_norm = mention.label // i.e., "Causal" or "Correlation"
+      relation_modifier = getModifier(mention) // prob none
+
+      effect <- mention.asInstanceOf[EidosEventMention].eidosArguments("effect")
+      factor_b_info = EntityInfo(effect, groundAs)
+
+      location = "" // I could try here..?
+      time = ""
+      evidence = removeTabAndNewline(mention.odinMention.sentenceObj.getSentenceText.trim)
+
+      row = source + "\t" + system + "\t" + sentence_id + "\t" +
+        factor_a_info.toTSV + "\t" +
+        relation_txt + "\t" + relation_norm + "\t" + relation_modifier + "\t" +
+        factor_b_info.toTSV + "\t" +
+        location + "\t" + time + "\t" + evidence + "\t" + factor_a_info.groundingToTSV + "\t" +
+        factor_b_info.groundingToTSV + "\n"
+    } pw.print(row)
+  }
+}
+
+
+
+
+
+
