+++ src/main/scala/org/clulab/wm/eidos/exporters/JSONLDExporter.scala
@@ -1,0 +1,21 @@
+package org.clulab.wm.eidos.exporters
+
+import org.clulab.serialization.json.stringify
+import org.clulab.wm.eidos.EidosSystem
+import org.clulab.wm.eidos.document.AnnotatedDocument
+import org.clulab.wm.eidos.serialization.json.JLDCorpus
+import org.clulab.wm.eidos.utils.Closer.AutoCloser
+import org.clulab.wm.eidos.utils.FileUtils
+
+
+case class JSONLDExporter(filename: String, reader: EidosSystem) extends Exporter {
+
+  override def export(annotatedDocuments: Seq[AnnotatedDocument]): Unit = {
+    FileUtils.printWriterFromFile(filename).autoClose { pw =>
+      val corpus = new JLDCorpus(annotatedDocuments)
+      val mentionsJSONLD = corpus.serialize()
+
+      pw.println(stringify(mentionsJSONLD, pretty = true))
+    }
+  }
+}
