+++ src/main/scala/org/clulab/wm/eidos/exporters/SerializedExporter.scala
@@ -1,0 +1,23 @@
+package org.clulab.wm.eidos.exporters
+
+import java.io.File
+
+import org.clulab.odin.Mention
+import org.clulab.utils.Serializer
+import org.clulab.wm.eidos.document.AnnotatedDocument
+
+case class SerializedExporter(filename: String) extends Exporter {
+
+  override def export(annotatedDocuments: Seq[AnnotatedDocument]): Unit = {
+    val odinMentions = annotatedDocuments.flatMap(ad => ad.odinMentions)
+    Serializer.save[SerializedMentions](new SerializedMentions(odinMentions), filename + ".serialized")
+  }
+}
+
+// Helper Class to facilitate serializing the mentions
+@SerialVersionUID(1L)
+class SerializedMentions(val mentions: Seq[Mention]) extends Serializable {}
+object SerializedMentions {
+  def load(file: File): Seq[Mention] = Serializer.load[SerializedMentions](file).mentions
+  def load(filename: String): Seq[Mention] = Serializer.load[SerializedMentions](filename).mentions
+}
